rag-setup/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
asyncpg==0.29.0

# Vector search
pgvector==0.2.4

# Web client
httpx==0.25.2
requests==2.31.0

# File handling
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.5.1
